API and move visualization
