+++ pypesto/visualize/__init__.py
@@ -26,6 +26,10 @@ from .optimization_stats import (
 )
 from .optimizer_convergence import optimizer_convergence
 from .optimizer_history import optimizer_history, optimizer_history_lowlevel
+from .ordinal_categories import (
+    plot_categories_from_inner_result,
+    plot_categories_from_pypesto_result,
+)
 from .parameters import (
     optimization_scatter,
     parameter_hist,
